requirements.txt
torch>=2.2
torchvision>=0.17
numpy>=1.26
pillow>=10.0
opencv-python-headless>=4.9
matplotlib>=3.8
scikit-learn>=1.4
tqdm>=4.66
fastapi>=0.111
uvicorn[standard]>=0.30
python-multipart>=0.0.9
requests>=2.32
streamlit>=1.36
